+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 Changelog for *Blade* project
 
+## blade-graphics-0.8.1 (TBD)
+
+- new API for enumerating device availability
+
 ## blade-graphics-0.8, blade-util-0.3, blade-egui-0.7, blade-particle-0.1, blade-asset-0.2.1, blade-engine-0.1 (26 Mar 2026)
 
 - examples: